requirements.txt
anthropic
yfinance
streamlit
pandas
openpyxl
xlsxwriter
requests
beautifulsoup4
plotly
python-dotenv
